Версия 3.16.3: тип данных для рейки в UIS изменён на double. Completed 21.04.21 21.00
- в InjectorUisTest, UisTestManager и в CustomTestUisDialogController сделаны необходимые изменения
- дополнительно:
 --добавлена форсунка CAT 3116 и изменен display_order в testNamesUIS.csv
 --выключено кодирование для 0445110720,Bosch
 --для форсунок Bosch уменьшен coding_range:
  -> взяты максимальные значения из оригинальных кодов (с фото форсунок) 0445117021, 0445117022, 0445117076, 0986435413 (копии 0445117021).
    --- Maximum Load 2.2 -> 0.9
    --- Emission Point 1.6 -> 0.8
  -> уменьшено по аналогии с предыдущими, фото не искал, были жалобы: 0445117028, 0445117029, 0445117070, 0986435431 (это копии 0445117029)
  -> взяты максимальные значения из оригинальных кодов (с фото форсунок): 0445116007, 0445116008, 0445116014, 0445116015, 0445116022, 0445116023, 0445116047, 0986435357, 0986435361, 0986435379 (это копии 0445116015)
    --- Emission Point 1.2 -> 1

*** в DelphiC3ICoding найден косяк в appendCheckSum(String binary): в for (int i = 93; i > 0; i--) должно быть i >= 0, но работает, т.к. там первый бит у binary всегда 0.

+++ src/main/java/fi/stardex/sisu/measurement/UisTestManager.java
@@ -370,7 +370,7 @@ public class UisTestManager implements TestManager {
             button.setDefaultButton(true);
             button.setOnAction(event -> start());
         }
-        int rackPosition = ((InjectorUisTest) (mainSectionUisModel.injectorTestProperty().get())).getRackPosition();
+        double rackPosition = ((InjectorUisTest) (mainSectionUisModel.injectorTestProperty().get())).getRackPosition();
         ((Button)settingsAlert.getDialogPane().lookupButton(ButtonType.APPLY)).textProperty().setValue(applyButton.get());
         settingsAlert.setContentText(alertString.get() + " " + rackPosition + " mm");
         settingsAlert.show();
